refactor: restructure SDMX-related modules and classes for improved organization and clarity
