+++ statgpt/common/data/proxy/sdmx_schemas/structure_message.py
@@ -1,12 +1,10 @@
 from pydantic import BaseModel, Field
 from sdmx.message import StructureMessage
 from sdmx.model.common import CubeRegion, DimensionComponent
-from sdmx.model.v21 import ContentConstraint, DataStructureDefinition, MemberSelection, MemberValue
+from sdmx.model.v21 import ContentConstraint, MemberSelection, MemberValue
 
-from statgpt.common.data.common.sdmx_schemas import (
+from statgpt.common.data.base.sdmx_schemas import (
     Sdmx30AnnotationModel,
-    Sdmx30DataComponentFilter,
-    build_availability_filters,
     to_content_constraint,
     to_structure_message,
 )
@@ -78,62 +76,6 @@ class ProxyAvailabilityData(BaseModel):
     data_constraints: list[ProxyDataConstraint] = Field(
         alias="dataConstraints", default_factory=list
     )
-
-
-class ProxyDataComponentFilter(Sdmx30DataComponentFilter):
-    pass
-
-
-class ProxyAvailabilityRequestBody(BaseModel):
-    """A request body in the JSON format for the Proxy SDMX 3.0 API."""
-
-    filters: list[ProxyDataComponentFilter] | None = Field(default=None)
-    key: str | None = Field(default=None)
-    component_id: str | None = Field(default="*")
-    # ~~~ Not used: ~~~
-
-    # updated_after: str
-    # references: str
-    # mode: str
-    # timestampTo: datetime
-    # keys: list[str]
-    # skipDeleted: bool
-    # dimensionAtObservation: str
-
-    @classmethod
-    def get_from(
-        cls,
-        key: dict[str, list[str]] | None,
-        params: dict[str, str] | None,
-        dsd: DataStructureDefinition | None,
-    ) -> "ProxyAvailabilityRequestBody":
-        return cls(
-            filters=build_availability_filters(ProxyDataComponentFilter, key, params),
-            key=cls._build_key_segment(key=key, dsd=dsd),
-        )
-
-    @classmethod
-    def _build_key_segment(
-        cls,
-        *,
-        key: dict[str, list[str]] | None,
-        dsd: DataStructureDefinition | None,
-    ) -> str:
-        if not key or not dsd:
-            return "*"
-        dim_ids = [
-            dim.id
-            for dim in dsd.dimensions.components
-            if not getattr(dim, "is_time_dimension", False) and dim.id != "TIME_PERIOD"
-        ]
-        parts = []
-        for dim_id in dim_ids:
-            values = key.get(dim_id)
-            if not values:
-                parts.append("")
-            else:
-                parts.append("+".join(values))
-        return ".".join(parts) or "*"
 
 
 class ProxyAvailabilityResponseBody(BaseModel):
