+++ statgpt/common/data/quanthub/sdmx_schemas/v30/__init__.py
@@ -1,7 +1,6 @@
 from .data_message import Attribute, QhDataMessage
 from .structure_message import (
     QhAnnotation,
-    QhAvailabilityRequestBody,
     QhAvailabilityResponseBody,
     QhDataflowMessage,
 )
