+++ statgpt/common/data/quanthub/v21/datasource.py
@@ -9,13 +9,18 @@ from sdmx.model.v21 import DataflowDefinition as DataFlow
 from statgpt.common.auth.auth_context import AuthContext
 from statgpt.common.config import multiline_logger as logger
 from statgpt.common.data.base import DataSourceType
-from statgpt.common.data.common import SdmxAugmentedDataSourceHandler
 from statgpt.common.data.quanthub.config import QuanthubDataSetConfig, QuanthubSdmxDataSourceConfig
 from statgpt.common.data.quanthub.v21.dataset import QuanthubSdmx21DataSet
 from statgpt.common.data.sdmx.v21.attribute import Sdmx21Attribute
-from statgpt.common.data.sdmx.v21.dataset import SdmxOfflineDataSet
+from statgpt.common.data.sdmx.v21.attributes_creator import Sdmx21AttributesCreator
+from statgpt.common.data.sdmx.v21.dataflow_loader import DataflowLoader
+from statgpt.common.data.sdmx.v21.dataset import InvalidConfigurationError, SdmxOfflineDataSet
+from statgpt.common.data.sdmx.v21.datasource import Sdmx21DataSourceHandler
+from statgpt.common.data.sdmx.v21.dimensions_creator import DimensionsCreator
 from statgpt.common.data.sdmx.v21.ratelimiter import SdmxRateLimiterFactory
 from statgpt.common.data.sdmx.v21.schemas import StructureMessage21, Urn
+from statgpt.common.data.sdmx.v21.urn_utils import is_wildcarded_version, lookup_urn
+from statgpt.common.schemas.dataset import Status
 from statgpt.common.settings.sdmx import quanthub_settings
 from statgpt.common.utils import Cache
 from statgpt.common.utils.timer import debug_timer
@@ -23,9 +28,8 @@ from statgpt.common.utils.timer import debug_timer
 from .qh_sdmx_client import AsyncQuanthubClient
 
 
-class QuanthubSdmx21DataSourceHandler(SdmxAugmentedDataSourceHandler):
+class QuanthubSdmx21DataSourceHandler(Sdmx21DataSourceHandler):
 
-    # TEMP fix:
     _dataset_cache: Cache[QuanthubSdmx21DataSet] = Cache(ttl=quanthub_settings.dataset_cache_ttl)
 
     def __init__(self, config: QuanthubSdmxDataSourceConfig):
@@ -59,13 +63,14 @@ class QuanthubSdmx21DataSourceHandler(SdmxAugmentedDataSourceHandler):
                 )
                 return True
             except HTTPStatusError as e:
-                # availability endpoint returns 400 with NotFound instead of 403
-                # treat 400 as Forbidden as well
                 if e.response.status_code in [403, 400]:
-                    # 403 means user doesn't have access to dataset
                     return False
                 else:
                     raise
+
+    def _should_use_cache(self, allow_cached: bool) -> bool:
+        auth_enabled = getattr(self._config, "auth_enabled", False)
+        return allow_cached and not auth_enabled
 
     async def _load_extra_dataset_data(
         self, sdmx_client: Any, urn: Urn, structure_message: StructureMessage21 | None = None
@@ -114,11 +119,10 @@ class QuanthubSdmx21DataSourceHandler(SdmxAugmentedDataSourceHandler):
         attributes: list[Sdmx21Attribute],
         extra_data: Mapping[str, Any],
     ) -> QuanthubSdmx21DataSet:
-        config = dataset_config
         return QuanthubSdmx21DataSet(
             entity_id=entity_id,
             title=title,
-            config=config,
+            config=dataset_config,
             handler=self,
             dataflow=dataflow,
             locale=self._config.locale,
@@ -127,6 +131,117 @@ class QuanthubSdmx21DataSourceHandler(SdmxAugmentedDataSourceHandler):
             attribute_values=extra_data.get("attribute_values", {}),
             annotations=extra_data.get("annotations", []),
         )
+
+    async def _get_dataset(  # type: ignore[override]
+        self,
+        entity_id: uuid.UUID,
+        title: str,
+        config: dict,
+        auth_context: AuthContext,
+        allow_offline: bool = False,
+        allow_cached: bool = False,
+    ) -> QuanthubSdmx21DataSet | SdmxOfflineDataSet:
+        dataset_config = self.parse_data_set_config(config)
+
+        if self._should_use_cache(allow_cached):
+            if ds := self._dataset_cache.get(str(entity_id)):
+                logger.debug(
+                    f"Returning cached dataset(id={entity_id}, urn={dataset_config.urn!r})"
+                )
+                return ds
+
+        logger.info(f"Loading dataset urn={dataset_config.urn!r}.")
+        sdmx_client = await self.create_sdmx_client(auth_context)
+
+        try:
+            urn = Urn(
+                agency_id=dataset_config.urn.agency_id,
+                resource_id=dataset_config.urn.resource_id,
+                version=dataset_config.urn.version,
+            )
+            dataflow_loader = DataflowLoader(sdmx_client)
+            urn, structure_message = await dataflow_loader.load_structure_message(urn, mode="full")
+        except Exception:
+            if allow_offline:
+                msg = f"Failed to load the dataflow or its associated structures. urn={dataset_config.urn!r}"
+                logger.exception(msg)
+                status = Status(status='offline', details=msg)
+                return SdmxOfflineDataSet(entity_id, title, dataset_config, self, status)
+            raise
+
+        try:
+            dimensions_creator = DimensionsCreator(
+                structure_message, urn, self._config.locale, dataset_config.get_dimension_aliases()
+            )
+            dimensions = await dimensions_creator.create_dimensions()
+        except Exception:
+            if allow_offline:
+                msg = "Failed to create dimensions from the loaded structure message."
+                logger.exception(msg)
+                status = Status(status='offline', details=msg)
+                return SdmxOfflineDataSet(entity_id, title, dataset_config, self, status)
+            raise
+
+        try:
+            attributes_creator = Sdmx21AttributesCreator(
+                structure_message, urn, self._config.locale
+            )
+            attributes = await attributes_creator.create_attributes()
+        except Exception:
+            if allow_offline:
+                msg = "Failed to create attributes from the loaded structure message."
+                logger.exception(msg)
+                status = Status(status='offline', details=msg)
+                return SdmxOfflineDataSet(entity_id, title, dataset_config, self, status)
+            raise
+
+        try:
+            extra_data = await self._load_extra_dataset_data(
+                sdmx_client=sdmx_client, urn=urn, structure_message=structure_message
+            )
+        except Exception:
+            if allow_offline:
+                msg = "Failed to load additional dataset metadata."
+                logger.exception(f"{msg}. See exception details below.")
+                status = Status(status='offline', details=msg)
+                return SdmxOfflineDataSet(entity_id, title, dataset_config, self, status)
+            raise
+
+        try:
+            dataflow = structure_message.dataflow[urn]
+            if is_wildcarded_version(dataflow.structure.version):
+                dataflow.structure = lookup_urn(
+                    structure_message.structure, Urn.for_artifact(dataflow.structure)
+                )
+            result = self._build_dataset(
+                entity_id=entity_id,
+                title=title,
+                dataset_config=dataset_config,
+                dataflow=dataflow,
+                dimensions=dimensions,
+                attributes=attributes,
+                extra_data=extra_data,
+            )
+        except InvalidConfigurationError as e:
+            if allow_offline:
+                msg = f"Invalid dataset(urn={dataset_config.urn!r}) configuration: {e}"
+                logger.warning(msg)
+                status = Status(status='invalid_config', details=msg)
+                return SdmxOfflineDataSet(entity_id, title, dataset_config, self, status)
+            raise
+        except Exception:
+            if allow_offline:
+                msg = "Failed to create dataset class."
+                logger.exception(f"{msg}. See exception details below.")
+                status = Status(status='offline', details=msg)
+                return SdmxOfflineDataSet(entity_id, title, dataset_config, self, status)
+            raise
+
+        if self._should_use_cache(allow_cached):
+            self._dataset_cache.set(str(entity_id), result)
+            logger.info(f"Cached dataset(id={entity_id}, urn={dataset_config.urn!r}).")
+
+        return result
 
     async def get_dataset(
         self,
