+++ statgpt/common/data/sdmx/v21/attributes_creator.py
@@ -9,6 +9,7 @@ from statgpt.common.data.sdmx.v21.attribute import (
     Sdmx21StringAttribute,
 )
 from statgpt.common.data.sdmx.v21.schemas import StructureMessage21, Urn
+from statgpt.common.data.sdmx.v21.urn_utils import lookup_urn
 
 
 class Sdmx21AttributesCreator:
@@ -23,7 +24,8 @@ class Sdmx21AttributesCreator:
 
     @property
     def _dsd(self) -> DataStructureDefinition:
-        return self._structure_message.dataflow[self._urn].structure
+        dsd_ref = self._structure_message.dataflow[self._urn].structure
+        return lookup_urn(self._structure_message.structure, Urn.for_artifact(dsd_ref))
 
     @property
     def _sdmx_attributes(self) -> list[common.DataAttribute]:
@@ -44,7 +46,7 @@ class Sdmx21AttributesCreator:
             raise ValueError(f"{attribute=} does not contain required concept_identity.parent")
 
         urn = Urn.for_artifact(concept_identity.parent)  # type: ignore[arg-type]
-        schema = self._structure_message.concept_scheme[urn]
+        schema = lookup_urn(self._structure_message.concept_scheme, urn)
         return schema.items[concept_identity.id]
 
     def _create_attribute_from_concept(self, attribute: common.DataAttribute) -> Sdmx21Attribute:
@@ -109,7 +111,7 @@ class Sdmx21AttributesCreator:
 
         # Actually, `code_list_ref` is a valid code list, but it does not contain the data,
         # since we are loading the codelist separately.
-        code_list = self._structure_message.codelist[Urn.for_artifact(code_list_ref)]
+        code_list = lookup_urn(self._structure_message.codelist, Urn.for_artifact(code_list_ref))
 
         indexed_codelist = InMemoryCodeList(code_list, self._locale)
         return self._create_sdmx_code_list_attribute(attribute, name, description, indexed_codelist)
